+++ src/main/java/com/devconmx/nutrilud_backend/repository/T_recordatoriosRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.devconmx.nutrilud_backend.model.T_recordatoriosVO;
 
 @Repository
-public interface T_recordatoriosRepository extends JpaRepository<T_recordatoriosVO, Long> {
+public interface T_recordatoriosRepository extends JpaRepository<T_recordatoriosVO, Integer> {
 
 }
