+++ src/main/java/com/devconmx/nutrilud_backend/repository/Tdatos_consultasRepository.java
@@ -8,6 +8,6 @@ import org.springframework.stereotype.Repository;
 import com.devconmx.nutrilud_backend.model.Tdatos_consultasVO;
 
 @Repository
-public interface Tdatos_consultasRepository extends JpaRepository<Tdatos_consultasVO, Long> {
+public interface Tdatos_consultasRepository extends JpaRepository<Tdatos_consultasVO, Integer> {
     List<Tdatos_consultasVO> findByNutriologoVO(Integer id);
 }
