+++ src/main/java/com/devconmx/nutrilud_backend/repository/TrolsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.devconmx.nutrilud_backend.model.TrolsVO;
 
 @Repository
-public interface TrolsRepository extends JpaRepository<TrolsVO, Long> {
+public interface TrolsRepository extends JpaRepository<TrolsVO, Integer> {
 
 }
