+++ src/main/java/com/devconmx/nutrilud_backend/repository/Tusuario_adminsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.devconmx.nutrilud_backend.model.Tusuario_adminsVO;
 
 @Repository
-public interface Tusuario_adminsRepository extends JpaRepository<Tusuario_adminsVO, Long> {
+public interface Tusuario_adminsRepository extends JpaRepository<Tusuario_adminsVO, Integer> {
 
 }
